Small after review fixes: Change structure.sql to schema.rb, Small migration and routes updates

+++ config/routes.rb
@@ -11,6 +11,6 @@ Rails.application.routes.draw do
 
   resources :submissions, path: 'podania', only: [:index, :show, :destroy] do
     post :submit, path: 'podat'
-    delete :destroy_all, path: 'zmazat', on: :collection
+    delete :destroy_all, path: 'zmazat', on: :collection if Rails.env.development?
   end
 end
